+++ packages/11ty/_includes/web-components/image-sequence/index.js
@@ -3,15 +3,24 @@ import { createRef, ref } from 'lit/directives/ref.js'
 
 import { imageSequenceStyles } from './styles.js'
 
-// TODO: Ensure buffer size is > then performRotation increment (set bufferSize larger, make prefetch return a promise that we .then() )
-// TODO: Raise and set an error message on the status-overlay component if fetches error out and cursor: not-allowed
 
+/**
+ * @class ImageSequence
+ * @description A reactive Lit element for showing and interacting with a sequence of images.
+ *
+ * @todo on fetch error raise an error message on the status-overlay component
+ * and set cursor: not-allowed
+ */
 class ImageSequence extends LitElement {
 
   static styles = [ imageSequenceStyles ]
 
   static properties = {
-    bufferSize: {
+    animationFrame: { 
+      type: Number,
+      state: true,
+    },
+    bufferSize: { 
       type: Number,
       state: true,
     },
@@ -54,38 +63,45 @@ class ImageSequence extends LitElement {
   canvasRef = createRef()
 
   /**
-   * @property bufferReady
+   * @private
+   * @property #bufferWindow
    * 
-   * Returns an array of indexes to buffer
-   * 
-   **/ 
-  get indexesToBuffer() {
+   * @returns an array of indexes included in `bufferSize` given `index`
+   */ 
+  get #bufferWindow() {
     // NB: Calculated one length higher to protect against numerical under run
-    return Array(this.bufferSize).fill(0).map( (_,i) => ( this.images.length + this.index + i - Math.round(this.bufferSize/2) ) % this.images.length )
+    const imageCount = this.images.length
+    const windowStart = imageCount + this.index - Math.round(this.bufferSize/2)
+    
+    return Array(this.bufferSize)
+      .fill(0)
+      .map((_, i) => ((windowStart + i) % imageCount))
   }
 
   /**
-   * @property bufferReady
+   * @private
+   * @property #bufferReady
    * 
-   * Returns true if the buffer is loaded ahead and behind of `index`
-   * 
-   **/ 
-  get bufferReady() {
-    return this.images.filter( (img,j) => this.indexesToBuffer.includes(j) && img !== null ).length === this.bufferSize
+   * @returns true if the buffer is loaded ahead and behind of `index`
+   */
+  get #bufferReady() {
+    return this.images.filter((img, j) => this.#bufferWindow.includes(j) && img !== null ).length === this.bufferSize
   }
 
   /**
-   * @property buffered
-   * 
-   * Returns true if the buffer is loaded ahead and behind of `index`
-   * 
-   **/ 
+   * @property bufferedPct
+   * @returns percent of the buffer that is not-null
+   */
   get bufferedPct() {
-    return Math.floor( this.images.filter( (img,j) => this.indexesToBuffer.includes(j) && img !== null ).length / this.bufferSize * 100 ) 
+    return Math.floor( this.images.filter((img, j) => this.#bufferWindow.includes(j) && img !== null).length / this.bufferSize * 100 )
   }
 
+  /**
+   * @property someImagesLoaded
+   * @returns true if there is at least one image loaded
+   */
   get someImagesLoaded() {
-    return this.images.some( i => i !== null )
+    return this.images.some((image) => image !== null)
   }
 
   /**
@@ -93,46 +109,52 @@ class ImageSequence extends LitElement {
    * @param url {string} - image URL to fetch
    * @param seqIndex {Number} - index to store this image 
    * 
-   * Fetches `url`, converts it into a blob and stores the image data, optionally drawing to the canvas.
+   * Fetches `url`, converts it into a blob and stores the image data in `seqIndex`.
    * 
-   * The in-flight fetch is stored in requests[seqIndex] for cancellation and request deduplication.
-   **/
-  #fetchImage(url,seqIndex) {
-    const req = new Request(url)
+   * The in-flight fetch is stored in this.requests for cancellation and request deduplication, nulled on completion.
+   * 
+   * @returns {Promise} fetch resposne
+   */
+  #fetchImage(url, seqIndex) {
+    if (this.requests[seqIndex]) return
 
-    if (this.requests[seqIndex]) {
-      return
-    }
+    const request = new Request(url)
 
-    this.requests[seqIndex] = fetch(req)
-      .then( (resp) => resp.blob() )
-      .then( (blob) => window.createImageBitmap(blob) )
-      .then( (bmp) => {
+    const response = fetch(request)
+      .then((response) => response.blob())
+      .then((blob) => window.createImageBitmap(blob))
+      .then((bmp) => {
         if (this.intrinsicHeight === 0) {
           this.intrinsicHeight = bmp.height
           this.intrinsicWidth = bmp.width   
         }
         this.images[seqIndex] = bmp
-        // Draw if the user hasn't already gone past this index
-        if (this.index===seqIndex) {
-          this.#paintCanvas(bmp)
-          return        
-        } 
       })
-      .then( () => {
+      .then(() => {
         this.requests[seqIndex] = null
         this.requestUpdate()
       })
-      .catch( (err) => {
-        console.error(err)
+      .catch((error) => {
+        console.error(error)
       })
+
+    this.requests[seqIndex] = response
+
+    return response
   }
 
+  /**
+   * @function connectedCallback
+   * 
+   * `lit` lifecycle method fired the first time the element is connected to the document
+   * 
+   * Used to register our visibility IntersectionObserver
+   */
   connectedCallback() {
     super.connectedCallback()
 
-    const callback = (entries,observer) => {
-      entries.forEach( (entry) => {
+    const callback = (entries, observer) => {
+      entries.forEach((entry) => {
         if (entry.isIntersecting) {
           this.visible = true
           observer.disconnect()
@@ -163,7 +185,7 @@ class ImageSequence extends LitElement {
 
     // Internal state
     const pctToBuffer = 0.2
-    this.bufferSize = Math.ceil( this.imageUrls.length * pctToBuffer )
+    this.bufferSize = Math.ceil(this.imageUrls.length * pctToBuffer)
     this.blitting = null // null | animationFrameRequestId
     this.images = Array(this.imageUrls.length).fill(null) // Array< null | ImageBitmap >
     this.requests = Array(this.imageUrls.length).fill(null) // Array< null | Promise >
@@ -182,6 +204,8 @@ class ImageSequence extends LitElement {
   }
 
   /**
+   * @function debounce
+   * 
    * Returns a function, that, as long as it continues to be invoked, will not
    * be triggered. The function will be called after it stops being called for
    * N milliseconds. If `immediate` is passed, trigger the function on the
@@ -199,9 +223,10 @@ class ImageSequence extends LitElement {
    * @param {function} func - the function to execute
    * @param {integer} wait - the tine to wait in milliseconds
    * @param {boolean} immediate - whether to call the function immediately or at the end of the timeout
+   *
    * @returns
    */
-  debounce(func, wait = 250, immediate = false) {
+  debounce(fn, wait = 250, immediate = false) {
     let timeout = null
 
     return function () {
@@ -210,23 +235,30 @@ class ImageSequence extends LitElement {
       const later = function () {
         timeout = null
         if (!immediate) {
-          func.apply(context, args)
+          fn.apply(context, args)
         }
       }
       const callNow = immediate && !timeout
       clearTimeout(timeout)
       timeout = setTimeout(later, wait)
       if (callNow) {
-        func.apply(context, args)
+        fn.apply(context, args)
       }
     }
   }
 
+  /**
+   * @function handleMouseMove
+   * 
+   * @param {Event.buttons}
+   * @param {Event.clientX}
+   * 
+   * Sets interaction flag, hides overlays, and handles reversability check
+   */
   handleMouseMove({ buttons, clientX }) {
     if (buttons) {
       this.didInteract = true
 
-      this.hideOverlays()
       if (this.oldX) {
         const deltaX = clientX - this.oldX
         const deltaIndex = Math.floor(Math.log(Math.abs(deltaX))) || 1
@@ -246,12 +278,6 @@ class ImageSequence extends LitElement {
     }
   }
 
-  hideOverlays() {
-    this.querySelectorAll('.overlay').forEach((element) => {
-      element.classList.remove('visible')
-    })
-  }
-
   /**
    * @function nextImage
    * @param {Integer} n Number of steps between start index and end index
@@ -269,7 +295,7 @@ class ImageSequence extends LitElement {
    * @function #draw
    * 
    * Performs drawing operations against `this.context`
-   **/
+   */
   #draw(image) {
     this.context.drawImage(image,0,0)
   }
@@ -279,7 +305,7 @@ class ImageSequence extends LitElement {
    * @param {ImageBitmap} - image - image to paint 
    * 
    * Paints the `canvas` element with the image from this.index
-   **/
+   */
   #paintCanvas(image) {
     if (!this.canvasRef.value) {
       return
@@ -308,70 +334,86 @@ class ImageSequence extends LitElement {
    * @param changedProperties
    * 
    * `lit` lifecycle method for changed properties
-   * 
-   **/ 
+   */
   willUpdate(changedProperties) {
+    // Determine the animation indices, preload them, and then do the rotation 
     if (changedProperties.has('rotateToIndex') && this.rotateToIndex!==false) {
-      this.performRotation(this.rotateToIndex)
+      const frameCount = this.rotateToIndex - this.index + this.bufferSize / 2
+      const animationIndices = Array(frameCount).fill(0).map((_, i) => this.index + i + 1)
+      this.#preloadImages(animationIndices).then(this.animateRotation(this.rotateToIndex))
     }
+
+    // Draws `animationFrame` directly to canvas (for use )
+    if (changedProperties.has('animationFrame')) {
+      if (!this.animationFrame) { return }
+      if (this.images[this.animationFrame] === null) { return }
+
+      this.#paintCanvas(this.images[this.animationFrame])
+    } 
 
     if (changedProperties.has('index') && this.someImagesLoaded) {
-      this.#preloadImages()
-
-      if (this.bufferReady) {
-        this.#paintCanvas()      
-      }
+      this.#preloadImages().then(() => {
+        this.animateRotation(this.index)
+      })
     }
 
+    // Load enough to prepare for interaction
     if (changedProperties.has('visible') && !this.visible) {
-      this.#preloadImages()
+      this.#preloadImages().then(() => this.#paintCanvas())
     }
   }
 
   /**
    * @function #preloadImages
    * 
-   * Loads the k images behind and ahead of this.index
-   **/
-  #preloadImages() {
-    if (!this.images.some(i => i === null)) { return }
+   * Loads images of this.indexesToBuffer -- ahead and behind `this.index`
+   * 
+   * @returns {Promise} - Promise resolution for all preloads
+   */
+  #preloadImages(bufferWindow) {
+    if (!this.images.some(i => i === null)) return Promise.all([]);
 
-    // Really just making buffer counting ergonomic / readable here
-    // const indexesToBuf = Array(this.bufferSize).fill(0).map( (_,i) => ( this.images.length + this.index + i - Math.round(this.bufferSize/2) ) % this.images.length )
-    this.images.forEach( (image,i) => {
-      // Skip anything out of our range or already loaded
-      if ( !this.indexesToBuffer.includes(i) || image !== null ) {
-        return
-      } 
+    const indexesToLoad = bufferWindow ?? this.#bufferWindow
+    const imageRequests = this.images
+      .map((image, i) => {
+        // Skip anything out of our range or already loaded
+        if (!indexesToLoad.includes(i) || image !== null) return null;
+        const url = this.imageUrls[i]
+        return this.#fetchImage(url, i);
+      })
+      .filter((imageRequest) => imageRequest)
 
-      const url = this.imageUrls[i]
-      this.#fetchImage(url,i)
-    })
+    return Promise.all(imageRequests)
   }
 
   /**
    * Animates a rotation by stepping through images from the current index to the provided `newValue`
    */
-  performRotation(indexToMove) {
-    if (this.index === indexToMove) return
+  animateRotation(untilIndex) {
+    if (this.animationFrame === untilIndex) return
+
+    this.animationFrame = this.index
+
     const interval = setInterval(() => {
       /**
        * Set rotateToIndex to false when rotation is done and clear the interval
        */
-      if (this.index === indexToMove) {
+      if (this.animationFrame === untilIndex) {
+        this.index = untilIndex
         this.rotateToIndex = false
+        this.synchronizeSequenceInstances()
       }
       /**
        * Clear the interval if user has triggered another rotation
        */
-      if (this.rotateToIndex !== indexToMove) {
+      if (this.rotateToIndex !== untilIndex) {
         clearInterval(interval)
         return
       }
       /**
        * Step through images
        */
-      this.nextImage()
+      this.animationFrame += 1
     }, this.transition)
   }
 
@@ -386,7 +428,6 @@ class ImageSequence extends LitElement {
     this.index = newIndex
   }
 
-  // TODO: Consult quire team for expected behavior here
   synchronizeSequenceInstances() {
     clearTimeout(this.updateTimer)
     this.updateTimer = setTimeout(() => {
@@ -400,7 +441,6 @@ class ImageSequence extends LitElement {
   }
 
   render() {
-
     const descriptionOverlay = this.isInteractive ? 
       html`<div slot='overlay' class="overlay ${ this.didInteract === false ? 'visible' : '' }"><span class="description">
             <svg class="description__icon">
@@ -415,19 +455,18 @@ class ImageSequence extends LitElement {
           </span></div>` 
       : ''
 
-    return html`<div class="image-sequence ${ this.bufferReady ? '' : 'loading' } ${ this.isInteractive ? 'interactive' : '' }">
-                  <slot name="loading-overlay">
-                    <div slot="loading-overlay" class='${ this.bufferReady ? '' : 'visible'} loading overlay'>
-                      <div class="buffering-indicator">Loading Image Sequence&nbsp;(${ this.bufferedPct }%)...</div>
-                      </div>
-                  </slot>
-                  <canvas ${ref(this.canvasRef)} height="${this.intrinsicHeight}" width="${this.intrinsicWidth}" class="${ this.someImagesLoaded ? 'visible' : '' } ${ this.didInteract ? '' : 'fade-in' }" slot="images"></canvas>                
-                  <slot name="overlay">
-                    ${ descriptionOverlay }
-                  </slot>
-                </div>`
+    return html`<div class="image-sequence ${this.#bufferReady ? '' : 'loading'} ${this.isInteractive ? 'interactive' : ''}">
+        <slot name="loading-overlay">
+          <div slot="loading-overlay" class='${ this.#bufferReady ? '' : 'visible'} loading overlay'>
+            <div class="buffering-indicator">Loading Image Sequence&nbsp;(${ this.bufferedPct }%)...</div>
+            </div>
+        </slot>
+        <canvas ${ref(this.canvasRef)} height="${this.intrinsicHeight}" width="${this.intrinsicWidth}" class="${this.someImagesLoaded ? 'visible' : ''} ${this.didInteract ? '' : 'fade-in'}" slot="images"></canvas>
+        <slot name="overlay">
+          ${descriptionOverlay}
+        </slot>
+      </div>`
   }
-
 }
 
-customElements.define('q-image-sequence',ImageSequence)
+customElements.define('q-image-sequence', ImageSequence)